Update html, css and js files
