+++ desktop.css
@@ -229,9 +229,4 @@
     padding: 15px 38px 0 16px;
     height: 170px;
   }
-
-  .get-in-touch button {
-    margin-left: 0;
-    margin-bottom: 80px;
-  }
 }
