+++ index.html
@@ -137,7 +137,7 @@
 				</div>
 				<div class="get-in-touch">
 					<button type="button">
-						Get It Touch
+						Get In Touch
 					</button>
 				</div>
 			</form>
