+++ project-data.js
@@ -1,0 +1,52 @@
+const content = [
+    {
+
+        id: 1,
+        name: 'Tonic',
+        jobDetails: ['CANOPY', 'Back End Dev', '2015'],
+        content: 'A daily selection of privately personalized reads; no accounts or sign-ups required.',
+        description: "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum han printer took a galley of type and scrambled it 1960s with the releawn printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the relea'",
+        featuredImage: "./images/desktop-version/1.png",
+        technologies: ['html', 'css', 'javaScript'],
+        liveVersion: "url('https://pratap-kumar.github.io/')",
+        source: "url('https://github.com/PRATAP-KUMAR')"
+    },
+
+    {
+        id: 2,
+        name: 'Multi-post Stories',
+        jobDetails: ['Facebook', 'Full Stack Dev', '2015'],
+        content: 'Experimental content creation feature that allows users to add to an existing story over the course of a day without spamming their friends.',
+        description: "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum han printer took a galley of type and scrambled it 1960s with the releawn printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the relea'",
+        featuredImage: "./images/desktop-version/projects/2.png",
+        technologies: ['html','Ruby on rails', 'css', 'javaScript'],
+        liveVersion: "url('https://pratap-kumar.github.io/')",
+        source: "url('https://github.com/PRATAP-KUMAR')"
+    },
+
+    {
+        id: 3,
+        name: 'Facebook 360',
+        jobDetails: ['Facebook', 'Full Stack Dev', '2015'],
+        content: "Exploring the future of media in Facebook's first Virtual Reality app; a place to discover and enjoy 360 photos and videos on Gear VR.",
+        description: "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum han printer took a galley of type and scrambled it 1960s with the releawn printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the relea'",
+        featuredImage: "./images/desktop-version/projects/3.png",
+        technologies: ['html','Ruby on rails', 'css', 'javaScript'],
+        liveVersion: "url('https://pratap-kumar.github.io/')",
+        source: "url('https://github.com/PRATAP-KUMAR')"
+    },
+
+    {
+        id: 4,
+        name: 'Uber Navigation',
+        jobDetails: ['Facebook', 'Lead Developer', '2018'],
+        content: 'A smart assistant to make driving more safe, efficient, and fun by unlocking your most expensive computer: your car.',
+        description: "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum han printer took a galley of type and scrambled it 1960s with the releawn printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the relea'",
+        featuredImage: "./images/desktop-version/projects/4.png",
+        technologies: ['html','Ruby on rails', 'css', 'javaScript'],
+        liveVersion: "url('https://pratap-kumar.github.io/')",
+        source: "url('https://github.com/PRATAP-KUMAR')"
+    }
+]
+
+export default content;
